Extract firebase function into separate file
